Add MongoDB Atlas cluster creation documentation
This commit includes detailed documentation on how to create and deploy a MongoDB Atlas cluster. The documentation is accompanied by multiple images that illustrate each step of the process, from creating a new organization and project to loading sample datasets. The README.md file has also been added to provide an overview of the repository.